adicionado o owner da task

+++ app/routers/tasks.py
@@ -4,6 +4,7 @@ from ..database import get_db
 from fastapi import  Depends, status, HTTPException, APIRouter
 from sqlalchemy.orm import Session
 from typing import List
+from typing import Optional
 
 router = APIRouter(
     prefix="/tasks", #Define o prefixo da rota,
@@ -13,13 +14,13 @@ router = APIRouter(
 
 # Rota para criar uma nova tarefa
 @router.post("/",status_code=status.HTTP_201_CREATED, response_model=schemas.task)
-def create_new_task(task: schemas.CreateTask, db: Session = Depends(get_db), current_user: schemas.User = Depends(oauth2.get_current_user)):
-    return create_task(db, task)
+async def create_new_task(task: schemas.CreateTask, db: Session = Depends(get_db), current_user: schemas.User = Depends(oauth2.get_current_user)):
+    return create_task(db, task, current_user)
 
 # Rota para obter todas as tarefas cadastradas
 @router.get("/", response_model=List[schemas.task])
-def get_tasks(db: Session = Depends(get_db)):
-    return read_tasks(db)
+def get_tasks(db: Session = Depends(get_db), search: Optional[str] = ""):
+    return read_tasks(db, search)
 
 #Rota para ver uma tarefa desejada
 @router.get("/{id}", response_model=schemas.task)
@@ -34,7 +35,9 @@ def get_task(task_id: int, db: Session = Depends(get_db)):
 def update_task_status(task_id: int, status: bool, db: Session = Depends(get_db), current_user: schemas.User = Depends(oauth2.get_current_user)):
     db_task = db.query(models.Task).filter(models.Task.id == task_id).first()
     if not db_task:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Task not found")
+        raise HTTPException(status_code=404, detail="Task not found")
+    if db_task.owner_id != current_user.id:
+        raise HTTPException(status_code=401, detail="You are not the owner of this task")
     if status == True:
         db_task.status = 'Concluido'
     else:
@@ -49,4 +52,6 @@ def delete_task_by_id(task_id: int, db: Session = Depends(get_db), current_user:
     db_task = read_task(db, task_id)
     if db_task is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Task not found")
+    if db_task.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="You are not the owner of this task")
     return delete_task(db, task_id)
